feat: move app to /schools route with clean query URLs
Filter links now live at /schools?types=District instead of the
root, with the query string always built explicitly against the
/schools path so it never picks up a stray slash before the "?".
Root now client-redirects to /schools (preserving any query string)
since this is a static export with no server-side redirects.
